Add 2021 NDCH event RSVP to header
Our site does not have any information for the National Day of Civic
Hacking, and a simple solution is to add a link. Therefore, add a link
to the event Meetup page.

---
This seemed like an easier idea than creating a whole page. I also
didn't extend the blue border bottom because that's currently hard
coded.

To undo this add, we just make a git-revert after the event.