Extend contact form text area vertically (#59)
The text area was about the same height as the input boxes. This makes
it hard for users to write a message and to scroll through their
message. Fix this by making the height of the text area larger for
longer messages.

Additionally, remove the height: 100% so the content of the container
does not extend past the contact box.